requirements-dev.txt
requests>=2.31.0
PyQt6>=6.5.0
pytchat>=0.9.0
openai>=1.0.0
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0
python-dotenv>=1.0.0
psutil>=5.9.0
pyinstaller>=5.13.0
